+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Changed
+- Use `useQuery` and `useMutation` hooks in ProductContext and OrderFormProvider.
 
 ## [2.78.0] - 2019-12-05
 ### Added